requirements.txt
annotated-types  
anyio  
attrs  
blinker  
blis  
boto3  
botocore  
Brotli  
cachetools  
catalogue  
certifi  
cffi  
charset-normalizer  
clean-text  
click  
cloudpathlib  
colorama  
comtypes  
confection  
cryptography  
cymem  
decorator  
dill  
distro  
elevenlabs  
emoji  
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256  
exceptiongroup  
ffmpeg-python  
filelock  
Flask  
fsspec  
ftfy  
future  
google-api-core  
google-api-python-client  
google-auth  
google-auth-httplib2  
google-auth-oauthlib  
googleapis-common-protos  
greenlet  
gTTS  
h11  
httpcore  
httplib2  
httpx  
huggingface-hub  
idna  
ImageHash  
imageio  
imageio-ffmpeg  
instagrapi  
intel-openmp  
itsdangerous  
Jinja2  
jiter  
jmespath  
langcodes  
language_data  
lxml  
marisa-trie  
markdown-it-py  
MarkupSafe  
mdurl  
mkl  
moviepy  
mpmath  
msedge-selenium-tools  
multiprocess  
murmurhash  
mutagen  
networkx  
numpy  
oauthlib  
openai  
opencv-python  
outcome  
packaging  
pathos  
pillow  
playwright  
pox  
ppft  
praw  
prawcore  
preshed  
proglog  
proto-plus  
protobuf  
pyasn1  
pyasn1_modules  
pycparser  
pycryptodomex  
pydantic  
pydantic_core  
pyee  
PyExecJS  
Pygments  
pynput  
pyparsing  
pyperclip  
pypiwin32  
PySocks  
python-dateutil  
python-dotenv  
pyttsx3  
pytz  
PyWavelets  
pywin32  
PyYAML  
regex  
requests  
requests-oauthlib  
rich  
rsa  
s3transfer  
safetensors  
schedule  
scipy  
selenium  
shellingham  
six  
smart-open  
sniffio  
sortedcontainers  
spacy  
spacy-legacy  
spacy-loggers  
srsly  
sympy  
tbb  
thinc  
tiktok-uploader  
tokenizers  
toml  
tomlkit  
torch  
tqdm  
transformers  
translators  
trio  
trio-websocket  
typer  
typing_extensions  
Unidecode  
update-checker  
uritemplate  
urllib3  
wasabi  
wcwidth  
weasel  
webdriver-manager  
websocket-client  
websockets  
Werkzeug  
wrapt  
wsproto  
yt-dlp  